add animation button
