+++ android/app/build.gradle
@@ -55,8 +55,8 @@ android {
         targetSdkVersion 35
         // versionCode flutterVersionCode.toInteger()
         // versionName flutterVersionName
-        versionCode 27
-        versionName "1.3.1"
+        versionCode 28
+        versionName "2.0.0"
     }
 
     signingConfigs {
