+++ lib/pages/byPackage/package_info.dart
@@ -6,74 +6,65 @@ import 'package:permission_handler/permission_handler.dart';
 import 'package:archive/archive_io.dart';
 import 'package:http/http.dart' as http;
 import 'package:flutter_svg/flutter_svg.dart';
-import 'package:wired_test/pages/home_page.dart';
-import '../providers/auth_guard.dart';
-import '../utils/custom_app_bar.dart';
-import '../utils/custom_nav_bar.dart';
-import '../utils/functions.dart';
-import '../utils/side_nav_bar.dart';
-import 'cme/cme_tracker.dart';
-import 'download_confirm.dart';
-import 'menu/guestMenu.dart';
-import 'menu/menu.dart';
-import 'module_library.dart';
-import '../services/location_service.dart';
+import '../../providers/auth_guard.dart';
+import '../../utils/custom_app_bar.dart';
+import '../../utils/custom_nav_bar.dart';
+import '../../utils/functions.dart';
+import '../../utils/side_nav_bar.dart';
+import '../cme/cme_tracker.dart';
+import '../download_confirm.dart';
+import '../home_page.dart';
+import '../menu/guestMenu.dart';
+import '../menu/menu.dart';
+import '../module_library.dart';
 
-
-class ModuleInfo extends StatefulWidget {
-  final int moduleId;
-  final String moduleName;
-  final String moduleDescription;
+class PackageInfo extends StatefulWidget {
+  final int packageId;
+  final String packageName;
+  final String packageDescription;
   final String? downloadLink;
 
-  LocationService locationService = LocationService();
-
-  ModuleInfo({
-    required this.moduleId,
-    required this.moduleName,
-    required this.moduleDescription,
+  PackageInfo({
+    required this.packageId,
+    required this.packageName,
+    required this.packageDescription,
     this.downloadLink
   });
 
   @override
-  _ModuleInfoState createState() => _ModuleInfoState();
+  _PackageInfoState createState() => _PackageInfoState();
 }
 
-class Modules {
+class Packages {
   String? name;
   String? description;
   String? downloadLink;
-  List<String>? letters;
 
-  Modules({
+  Packages({
     this.name,
     this.description,
     this.downloadLink,
-    this.letters,
   });
 
-  Modules.fromJson(Map<String, dynamic> json)
+  Packages.fromJson(Map<String, dynamic> json)
       : name = json['name'] as String,
         description = json['description'] as String,
-        downloadLink = json['downloadLink'] as String,
-        letters = (json['letters'] as List<dynamic>?)?.map((e) => e as String).toList();
+        downloadLink = json['downloadLink'] as String;
 
 
   Map<String, dynamic> toJson() => {
     'name': name,
     'description': description,
     'downloadLink': downloadLink,
-    'letters': letters,
   };
 }
 
-class _ModuleInfoState extends State<ModuleInfo> {
-  late Future<Modules> futureModule;
-  late List<Modules> moduleData = [];
-  final GlobalKey _moduleNameKey = GlobalKey();
+class _PackageInfoState extends State<PackageInfo> {
+  late Future<Packages> futurePackage;
+  late List<Packages> packageData = [];
+  final GlobalKey _packageNameKey = GlobalKey();
   double topPadding = 0;
   bool _isLoading = false;
-  Map<String, double?>? _location;
 
   // Get Permissions
   Future<bool> checkAndRequestStoragePermission() async {
@@ -84,49 +75,38 @@ class _ModuleInfoState extends State<ModuleInfo> {
     return status.isGranted;
   }
 
-  Future<String> getStoragePath() async {
-
-    Directory? directory;
-
-    if (Platform.isAndroid) {
-      directory = await getExternalStorageDirectory(); // Android external storage
-    } else if (Platform.isIOS || Platform.isMacOS) {
-      directory = await getApplicationSupportDirectory(); // iOS/macOS safe location
-    } else if (Platform.isWindows || Platform.isLinux) {
-      directory = await getApplicationDocumentsDirectory(); // Windows/Linux
-    }
-
-    return directory?.path ?? "/default/path"; // Fallback path
-  }
-
-  // Download the Module
-  Future<void> downloadModule(String url, String fileName) async {
+  // Download the Package
+  Future<void> downloadPackage(String url, String fileName) async {
     bool hasPermission = await checkAndRequestStoragePermission();
     print("Has Permission: $hasPermission");
-
     if (true) {
-      final storagePath = await getStoragePath();
-      final modulesDirectoryPath = '$storagePath/modules';
-      final modulesDirectory = Directory(modulesDirectoryPath);
-
-      // Check if the individuals modules directory exists
-      if (!modulesDirectory.existsSync()) {
-        modulesDirectory.createSync(recursive: true);
-        print('Directory created: $modulesDirectoryPath');
+      final directory = await getExternalStorageDirectory(); // Get the External Storage Directory (Android)
+      if (directory == null) {
+        ScaffoldMessenger.of(context).showSnackBar(
+          SnackBar(content: Text('Unable to access storage directory')),
+        );
+        return;
       }
 
-      final modulesFilePath = '$modulesDirectoryPath/$fileName';
-      final file = File(modulesFilePath);
+      // Check if the packages directory exists
+      final packagesDirectoryPath = '${directory.path}/packages';
+      final packagesDirectory = Directory(packagesDirectoryPath);
+      if (!packagesDirectory.existsSync()) {
+        packagesDirectory.createSync(recursive: true);
+        print('Created directory: $packagesDirectoryPath');
+      }
+
+      final packagesFilePath = '$packagesDirectoryPath/$fileName';
+      final file = File(packagesFilePath);
 
       try {
         final response = await http.get(Uri.parse(url));
         await file.writeAsBytes(response.bodyBytes);
-
         ScaffoldMessenger.of(context).showSnackBar(
           SnackBar(content: Text('Downloaded $fileName')),
         );
-        print('Directory: $storagePath');
-        print('File Path: $modulesFilePath');
+        print('Directory: ${directory.path}');
+        print('File Path: $packagesFilePath');
 
         // Unzip the downloaded file
         final bytes = file.readAsBytesSync();
@@ -134,29 +114,29 @@ class _ModuleInfoState extends State<ModuleInfo> {
 
         for (var file in archive) {
           final filename = file.name;
-          final extractedFilePath = '$modulesDirectoryPath/$filename';
-          print('Processing file: $filename at path: $extractedFilePath');
+          final packagesFilePath = '${directory.path}/packages/$filename';
+          print('Processing file: $filename at path: $packagesFilePath');
 
           if (file.isFile) {
             final data = file.content as List<int>;
-            File(extractedFilePath)
+            File(packagesFilePath)
               ..createSync(recursive: true)
               ..writeAsBytesSync(data);
           } else {
-            Directory(extractedFilePath).createSync(recursive: true);
-            print('Directory created: $extractedFilePath');
+            Directory(packagesFilePath).createSync(recursive: true);
+            print('Directory created: $packagesFilePath');
           }
         }
 
         ScaffoldMessenger.of(context).showSnackBar(
           SnackBar(content: Text('Unzipped $fileName')),
         );
-        print('Unzipped to: $storagePath');
+        print('Unzipped to: ${directory.path}');
 
         // Delete the zip file
         try {
           await file.delete();
-          print('Zip file deleted: $modulesFilePath');
+          print('Zip file deleted: $packagesFilePath');
           ScaffoldMessenger.of(context).showSnackBar(
             SnackBar(content: Text('Unzipped and deleted $fileName')),
           );
@@ -184,16 +164,7 @@ class _ModuleInfoState extends State<ModuleInfo> {
     super.initState();
   }
 
-  Future<void> getLocationAndSaveDownload() async {
-    var location = await widget.locationService.getLocation(context);
-    setState(() {
-      _location = location;  // Store the location in state
-    });
-    await widget.locationService.saveDownload(widget.moduleId, location);
-  }
-  
-
-// Consider using AutoSizeText for the module name instead of RichText
+// Consider using AutoSizeText for the package name instead of RichText
 
   @override
   Widget build(BuildContext context) {
@@ -331,9 +302,10 @@ class _ModuleInfoState extends State<ModuleInfo> {
     return Column(
       children: [
         SizedBox(
-            height: baseSize * (isTablet(context) ? 0.03 : 0.03),
+          height: baseSize * (isTablet(context) ? 0.03 : 0.03),
         ),
-        // Module Description Container
+
+        // Package Description Container
         Flexible(
           flex: 6,
           child: Stack(
@@ -343,44 +315,44 @@ class _ModuleInfoState extends State<ModuleInfo> {
                 decoration: BoxDecoration(
                   color: Colors.transparent,
                 ),
-                child: SingleChildScrollView(
-                  child: Padding(
-                    padding: EdgeInsets.only(
-                      bottom: 50,
-                      left: 10,
-                      right: 10,
-                    ),
-                    child: RichText(
-                      textAlign: TextAlign.center,
-                      text: TextSpan(
-                        children: [
-                          TextSpan(
-                            text: '${widget.moduleName}\n',
-                            style: TextStyle(
-                              fontSize: baseSize * (isTablet(context) ? 0.06 : 0.065),
-                              fontWeight: FontWeight.w500,
-                              color: Color(0xFF0070C0),
+                  child: SingleChildScrollView(
+                    child: Padding(
+                      padding: EdgeInsets.only(
+                        bottom: 50,
+                        left: 10,
+                        right: 10,
+                      ),
+                      child: RichText(
+                        textAlign: TextAlign.center,
+                        text: TextSpan(
+                          children: [
+                            TextSpan(
+                              text: '${widget.packageName}\n',
+                              style: TextStyle(
+                                fontSize: baseSize * (isTablet(context) ? 0.06 : 0.065),
+                                fontWeight: FontWeight.w500,
+                                color: Color(0xFF0070C0),
+                              ),
                             ),
-                          ),
-                          WidgetSpan(
-                            child: SizedBox(
-                              height: baseSize * (isTablet(context) ? 0.08 : 0.08),
+                            WidgetSpan(
+                              child: SizedBox(
+                                height: baseSize * (isTablet(context) ? 0.08 : 0.08),
+                              ),
                             ),
-                          ),
-                          TextSpan(
-                            text: '${widget.moduleDescription}\n',
-                            style: TextStyle(
-                              fontSize: baseSize * (isTablet(context) ? 0.04 : 0.045),
-                              fontWeight: FontWeight.w500,
-                              color: Colors.black,
+                            TextSpan(
+                              text: '${widget.packageDescription}\n',
+                              style: TextStyle(
+                                fontSize: baseSize * (isTablet(context) ? 0.04 : 0.045),
+                                fontWeight: FontWeight.w500,
+                                color: Colors.black,
+                              ),
                             ),
-                          ),
-                        ],
+                          ],
+                        ),
                       ),
                     ),
                   ),
                 ),
-              ),
               // Container for gradient text fade
               Positioned(
                 bottom: 0,
@@ -409,104 +381,105 @@ class _ModuleInfoState extends State<ModuleInfo> {
             ],
           ),
         ),
-        // Download Button
-        Flexible(
-          flex: 1,
-          child: Padding(
-            padding: const EdgeInsets.only(top: 4),
-            child: GestureDetector(
-              onTap: _isLoading
-                  ? null
-                  : () async {
 
-                if (widget.downloadLink != null) {
-                  setState(() {
-                    _isLoading = true;
-                  });
+              // Download Button
+              Flexible(
+                flex: 1,
+                child: Padding(
+                  padding: const EdgeInsets.only(top: 4),
+                  child: GestureDetector(
+                    onTap: _isLoading
+                        ? null
+                        : () async {
+                      if (widget.downloadLink != null) {
+                        setState(() {
+                          _isLoading = true;
+                        });
 
-                  String fileName = "$widget.moduleName.zip";
-                  await downloadModule(widget.downloadLink!, fileName);
-                  await getLocationAndSaveDownload();
-                  setState(() {_isLoading = false;});
+                        String fileName = "$widget.packageName.zip";
+                        await downloadPackage(
+                            widget.downloadLink!, fileName);
 
-                  Navigator.push(context,
-                    MaterialPageRoute(
-                      builder: (context) => DownloadConfirm(
-                        moduleName: widget.moduleName
-                      )
-                    ),
-                  );
-                  
-                } else {
-                  ScaffoldMessenger.of(context)
-                      .showSnackBar(
-                    SnackBar(content: Text(
-                        'No download link found for ${widget
-                            .moduleName}')),
-                  );
-                }
-              },
-              child: Container(
-                width: baseSize * (isTablet(context) ? 0.5 : 0.55),
-                height: baseSize * (isTablet(context) ? 0.10 : 0.12),
-                decoration: BoxDecoration(
-                  gradient: const LinearGradient(
-                    colors: [
-                      Color(0xFF0070C0),
-                      Color(0xFF00C1FF),
-                      Color(0xFF0070C0),
-                    ], // Your gradient colors
-                    begin: Alignment.topCenter,
-                    end: Alignment.bottomCenter,
-                  ),
-                  borderRadius: BorderRadius.circular(35),
-                  boxShadow: [
-                    BoxShadow(
-                      color: Colors.black.withOpacity(
-                          0.5),
-                      spreadRadius: 1,
-                      blurRadius: 5,
-                      offset: const Offset(1,
-                          3), // changes position of shadow
-                    ),
-                  ],
-                ),
-                child: Center(
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    crossAxisAlignment: CrossAxisAlignment.center,
-                    children: [
-                      _isLoading
-                          ? CircularProgressIndicator(
-                        color: Colors.white,
-                        strokeWidth: 3,
-                      )
-                          : Text(
-                        "Download",
-                        style: TextStyle(
-                          fontSize: baseSize * (isTablet(context) ? 0.071 : 0.071),
-                          fontWeight: FontWeight.w500,
-                          color: Colors.white,
+                        setState(() {
+                          _isLoading = false;
+                        });
+                        print("Package Id: ${widget.packageId}");
+                        Navigator.push(context,
+                            MaterialPageRoute(
+                                builder: (context) =>
+                                    DownloadConfirm(
+                                        packageName: widget
+                                            .packageName)));
+                      } else {
+                        ScaffoldMessenger.of(context)
+                            .showSnackBar(
+                          SnackBar(content: Text(
+                              'No download link found for ${widget
+                                  .packageName}')),
+                        );
+                      }
+                    },
+                    child: Container(
+                      width: baseSize * (isTablet(context) ? 0.5 : 0.55),
+                      height: baseSize * (isTablet(context) ? 0.10 : 0.12),
+                      decoration: BoxDecoration(
+                        gradient: const LinearGradient(
+                          colors: [
+                            Color(0xFF0070C0),
+                            Color(0xFF00C1FF),
+                            Color(0xFF0070C0),
+                          ], // Your gradient colors
+                          begin: Alignment.topCenter,
+                          end: Alignment.bottomCenter,
+                        ),
+                        borderRadius: BorderRadius.circular(30),
+                        boxShadow: [
+                          BoxShadow(
+                            color: Colors.black.withOpacity(
+                                0.5),
+                            spreadRadius: 1,
+                            blurRadius: 5,
+                            offset: const Offset(1,
+                                3), // changes position of shadow
+                          ),
+                        ],
+                      ),
+                      child: Center(
+                        child: Row(
+                          mainAxisAlignment: MainAxisAlignment.center,
+                          crossAxisAlignment: CrossAxisAlignment.center,
+                          children: [
+                            _isLoading
+                                ? CircularProgressIndicator(
+                              color: Colors.white,
+                              strokeWidth: 3,
+                            )
+                                : Text(
+                              "Download",
+                              style: TextStyle(
+                                fontSize: baseSize * (isTablet(context) ? 0.071 : 0.071),
+                                fontWeight: FontWeight.w500,
+                                color: Colors.white,
+                              ),
+                            ),
+                            const SizedBox(width: 7,),
+                            SvgPicture.asset(
+                              'assets/icons/download_icon.svg',
+                              // height: 42,
+                              // width: 42,
+                              height: baseSize * (isTablet(context) ? 0.0675 : 0.0675),
+                              width: baseSize * (isTablet(context) ? 0.0675 : 0.0675),
+                            ),
+                          ],
                         ),
                       ),
-                      const SizedBox(width: 7,),
-                      SvgPicture.asset(
-                        'assets/icons/download_icon.svg',
-                        // height: 42,
-                        // width: 42,
-                        height: baseSize * (isTablet(context) ? 0.0675 : 0.0675),
-                        width: baseSize * (isTablet(context) ? 0.0675 : 0.0675),
-                      ),
-                    ],
+                    ),
                   ),
                 ),
-              ),
-            ),
-          ),
-        )
-      ],
-    );
-  }
+              )
+            ],
+          );
+        }
 
   Widget _buildLandscapeLayout(screenWidth, screenHeight, baseSize) {
     return Column(
@@ -515,56 +488,56 @@ class _ModuleInfoState extends State<ModuleInfo> {
           height: baseSize * (isTablet(context) ? 0.05 : 0.03),
         ),
 
-        // Module Description Container
+        // Package Description Container
         Flexible(
           flex: 6,
           child: Stack(
             children: [
-              Container(
-                height: baseSize * (isTablet(context) ? 0.65 : 0.65),
-                //width: 400,
-                decoration: BoxDecoration(
-                  color: Colors.transparent,
-                ),
-                child: SingleChildScrollView(
-                  child: Padding(
-                    padding: EdgeInsets.only(
-                      bottom: 50,
-                      left: 20,
-                      right: 20,
-                    ),
-                    child: RichText(
-                      textAlign: TextAlign.center,
-                      text: TextSpan(
-                        children: [
-                          TextSpan(
-                            text: '${widget.moduleName}\n',
-                            style: TextStyle(
-                              //fontSize: 32.0,
-                              fontSize: baseSize * (isTablet(context) ? 0.06 : 0.065),
-                              fontWeight: FontWeight.w500,
-                              color: Color(0xFF0070C0),
+                Container(
+                  height: baseSize * (isTablet(context) ? 0.65 : 0.65),
+                  //width: 400,
+                  decoration: BoxDecoration(
+                    color: Colors.transparent,
+                  ),
+                  child: SingleChildScrollView(
+                    child: Padding(
+                      padding: EdgeInsets.only(
+                        bottom: 50,
+                        left: 20,
+                        right: 20,
+                      ),
+                      child: RichText(
+                        textAlign: TextAlign.center,
+                        text: TextSpan(
+                          children: [
+                            TextSpan(
+                              text: '${widget.packageName}\n',
+                              style: TextStyle(
+                                //fontSize: 32.0,
+                                fontSize: baseSize * (isTablet(context) ? 0.06 : 0.065),
+                                fontWeight: FontWeight.w500,
+                                color: Color(0xFF0070C0),
+                              ),
                             ),
-                          ),
-                          WidgetSpan(
-                            child: SizedBox(
-                              height: baseSize * (isTablet(context) ? 0.08 : 0.08),
+                            WidgetSpan(
+                              child: SizedBox(
+                                height: baseSize * (isTablet(context) ? 0.08 : 0.08),
+                              ),
                             ),
-                          ),
-                          TextSpan(
-                            text: '${widget.moduleDescription}\n',
-                            style: TextStyle(
-                              fontSize: screenHeight * (isTablet(context) ? 0.04 : 0.045),
-                              fontWeight: FontWeight.w500,
-                              color: Colors.black,
+                            TextSpan(
+                              text: '${widget.packageDescription}\n',
+                              style: TextStyle(
+                                fontSize: screenHeight * (isTablet(context) ? 0.04 : 0.045),
+                                fontWeight: FontWeight.w500,
+                                color: Colors.black,
+                              ),
                             ),
-                          ),
-                        ],
+                          ],
+                        ),
                       ),
                     ),
                   ),
                 ),
-              ),
 
               // Container for gradient text fade
               Positioned(
@@ -594,7 +567,8 @@ class _ModuleInfoState extends State<ModuleInfo> {
             ],
           ),
         ),
-        // Download Button
+
+              // Download Button
         Flexible(
           flex: 1,
           child: Padding(
@@ -608,26 +582,26 @@ class _ModuleInfoState extends State<ModuleInfo> {
                     _isLoading = true;
                   });
 
-                  String fileName = "$widget.moduleName.zip";
-                  await downloadModule(widget.downloadLink!, fileName);
-                  await getLocationAndSaveDownload();
-                  setState(() {_isLoading = false;});
+                  String fileName = "$widget.packageName.zip";
+                  await downloadPackage(
+                      widget.downloadLink!, fileName);
 
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                      builder: (context) => DownloadConfirm(
-                        moduleName: widget.moduleName
-                      )
-                    )
-                  );
-                  print("Module Id: ${widget.moduleId}");
+                  setState(() {
+                    _isLoading = false;
+                  });
+
+                  Navigator.push(context,
+                      MaterialPageRoute(
+                          builder: (context) =>
+                              DownloadConfirm(
+                                  packageName: widget
+                                      .packageName)));
                 } else {
                   ScaffoldMessenger.of(context)
                       .showSnackBar(
                     SnackBar(content: Text(
                         'No download link found for ${widget
-                            .moduleName}')),
+                            .packageName}')),
                   );
                 }
               },
@@ -693,4 +667,3 @@ class _ModuleInfoState extends State<ModuleInfo> {
     );
   }
 }
-
